uv-index-website/backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.1
requests==2.32.3
psycopg2-binary==2.9.5
python-dotenv==1.0.1
xmltodict==0.13.0
Werkzeug==3.1.3
Jinja2==3.1.6 